[RESTEASY-1715] Adding a warning in case of users forgetting to close apache http client 4.3 engine + providing another engine without finalizer for users who are brave enough to use it ;-)
